+++ src/main.ts
@@ -1925,11 +1925,7 @@ export default class VimMotionsPlugin extends Plugin {
                 this.settings,
             );
         }
-        if (
-            vim &&
-            this.settings.enableTableNav &&
-            this.settings.tableWidgetMode === 'native'
-        ) {
+        if (vim && this.settings.tableWidgetMode === 'native') {
             this.uninstallTableCellMotions = applyTableCellMotions(
                 this.app,
                 vim,
@@ -2564,10 +2560,7 @@ export default class VimMotionsPlugin extends Plugin {
                 this.settings,
             );
         }
-        if (
-            this.settings.enableTableNav &&
-            this.settings.tableWidgetMode === 'native'
-        ) {
+        if (this.settings.tableWidgetMode === 'native') {
             this.uninstallTableCellMotions = applyTableCellMotions(
                 this.app,
                 vim,